refactor(app): Move shared table logic to DataTable component.
